refactor: merge application/domain into top-level domain/
Consolidates the duplicated "pure business rules" concept:
authPolicy, actionRules, saveRules, and studentAccess move out of
application/domain/ into concern-scoped domain/ subfolders (auth/,
action/, savedStudent/, student/), alongside the existing
domain/Company/ and domain/Student/ entity folders, which are
renamed to camelCase (company/, student/) for naming consistency
with the new ones.
