+++ src/application/repositories/studentRepository.ts
@@ -1,7 +1,7 @@
 import "server-only";
 
 import { Email } from "@/types/Email";
-import type { StudentYear } from "@/domain/Student/year";
+import type { StudentYear } from "@/domain/student/year";
 
 import prisma, { DbClient } from "./database";
