+++ src/application/services/savedStudentService.ts
@@ -3,10 +3,10 @@ import "server-only";
 import { Email } from "@/types/Email";
 import { HttpError } from "@/types/HttpError";
 import type { Stats } from "@/types/Stats";
+import { assertStudentCanBeSaved } from "@/domain/savedStudent/saveRules";
 import { getBoothActionName } from "@/edition/actions";
 import { ISEP_EMAIL_DOMAIN } from "@/utils/isepEmail";
 
-import { assertStudentCanBeSaved } from "../domain/saveRules";
 import {
   findCompanyById,
   findCompanyEmployee,
